requirements.txt
accelerate==0.26.1
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==22.3.2
Automat==20.2.0
backcall==0.2.0
bleach==6.1.0
blinker==1.7.0
blis==0.7.11
braceexpand==0.1.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2023.11.17
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.2
cloudpathlib==0.16.0
colorama==0.4.6
comm==0.1.2
confection==0.1.4
constantly==15.1.0
contexttimer==0.3.3
contourpy==1.2.0
cryptography==42.0.2
cycler==0.12.1
cymem==2.0.8
debugpy==1.6.4
decorator==5.1.1
decord==0.6.0
distlib==0.3.8
distro==1.9.0
einops==0.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==1.2.0
fairscale==0.4.4
fastapi==0.109.0
ffmpy==0.3.1
filelock==3.13.1
fonttools==4.47.2
frozenlist==1.4.1
-e git+https://github.com/lm-sys/FastChat.git@7003bbd1ef1f26cdace1b0d3aedae6d25a3f943c#egg=fschat
fsspec==2023.12.2
ftfy==6.1.3
git-lfs==1.6
gitdb==4.0.11
GitPython==3.1.41
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
gradio==4.37.2
gradio_client==0.6.1
grpcio==1.60.1
grpcio-status==1.60.1
h11==0.14.0
hkdf==0.0.3
httpcore==1.0.2
httpx==0.26.0
huggingface-hub==0.20.3
humanize==4.0.0
hyperlink==21.0.0
icecream==2.1.3
identify==2.5.33
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
importlib-resources==6.1.1
incremental==21.3.0
iopath==0.1.10
ipykernel==6.19.4
ipython==8.7.0
jedi==0.18.2
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==7.4.8
jupyter_core==5.1.1
kaggle==1.6.3
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.3
magic-wormhole==0.12.0
markdown-it-py==3.0.0
markdown2==2.4.12
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.10
nest-asyncio==1.5.6
networkx==3.2.1
nh3==0.2.15
nodeenv==1.8.0
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
openai==1.10.0
opencv-python-headless==4.5.5.64
opendatasets==0.1.22
orjson==3.9.12
packaging==22.0
pandas==2.2.0
parso==0.8.3
peft==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
plotly==5.18.0
portalocker==2.8.2
pre-commit==3.6.0
preshed==3.0.9
prompt-toolkit==3.0.36
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocoevalcap==1.2
pycocotools==2.0.7
pycparser==2.21
pydantic==1.10.14
pydantic_core==2.14.6
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.14.0
PyNaCl==1.5.0
pyOpenSSL==24.0.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-magic==0.4.27
python-multipart==0.0.6
python-slugify==8.0.1
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
referencing==0.32.1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
rsa==4.9
ruff==0.1.14
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.1.99
service-identity==21.1.0
shellingham==1.5.4
shortuuid==1.0.11
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spake2==0.8
srsly==2.4.8
stack-data==0.6.2
starlette==0.35.1
streamlit==1.30.0
svgwrite==1.4.3
sympy==1.12
tenacity==8.2.3
text-unidecode==1.3
thinc==8.2.2
tifffile==2023.12.9
tiktoken==0.5.2
timm==0.4.12
tokenizers==0.15.1
toml==0.10.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.1.2
torchvision==0.16.2
tornado==6.2
tqdm==4.66.1
traitlets==5.8.0
transformers==4.37.2
transformers-stream-generator==0.0.4
triton==2.1.0
Twisted==22.4.0
txaio==22.2.1
txtorcon==22.0.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.27.0
validators==0.22.0
virtualenv==20.25.0
wasabi==1.1.2
watchdog==3.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
weasel==0.3.4
webdataset==0.2.86
webencodings==0.5.1
websockets==11.0.3
yarl==1.9.4
zipp==3.17.0
zope.interface==5.4.0
